(WiiU) Port to wut toolchain
This commit is enough to compile RetroArch using wut and more modern
versions of devkitPPC. Among the requisite ports and bugfixes, the wiiu
code has been switched to use wut's headers (where possible).

There's a handful of headers where RA's are more precise, and these
have been put into the `wiiu/include/` hierarchy mirroring WUT's folder
structure, but with `ra_` prepended to the filename to avoid include
directory priority fighting.

Other changes are related to the header changes:
- update defines to use WUT's new definition
- replace custom integer types with standard ones (eg u8 -> uint8_t)
- remove some workarounds/hacks in favor of more precise handling
  in WUT.

Cut features that will be re-added later:
- Current rpx symbol names in the exception handler
- somemopt networking optimisations
- Core loading
- libfat support - always uses fs:/vol/external01

(WiiU) memory: handle being in the background properly

Quitting on Aroma works a little differently to HBL - we need to use a ProcUI
loop and go into the background before quitting. This patch prepares the MEM1
and bucket allocators so they won't crash when this happens.

(WiiU) aroma: fix conflict with FTPiiU plug-in

== DETAILS
The HID polling thread was written to be run on a single core, and
basically completely occupies it, not allowing any other threads to
run on it.

It turns out, this is a problem in Aroma, because other things can
be running on your cores. For instance, the FTPiiU server thread runs
on the same core as the HID polling thread, which explains why FTPiiU
stopped working when RetroArch was running.

The other problem, is that if FTPiiU is holding the IPC BufPool mutex
when RA is started, a deadlock will eventually happen because RA will
try to acquire the mutex and the FTPii thread, being suspended, won't
be able to release it.

So!

To fix it:
- add a 500-microsecond sleep to the HID polling thread
- update to WUT's HID headers, which makes some minor changes to
  structure member names